Create a CLI command to start chat
